+++ quantforge/__init__.py
@@ -314,6 +314,7 @@ from .dunn_test import dunn_test
 from .jonckheere import jonckheere_terpstra_test
 from .cochran_mcnemar import mcnemar_test, cochran_q_test
 from .fisher_exact import fisher_exact_test
+from .deming import deming_regression, orthogonal_regression
 from .kalman_beta import kalman_regression_beta
 from .theil_sen import theil_sen
 from .robust_stats import (
@@ -873,7 +874,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "1.889.0"
+__version__ = "1.890.0"
 
 __all__ = [
     "OptionType",
@@ -1354,6 +1355,8 @@ __all__ = [
     "mcnemar_test",
     "cochran_q_test",
     "fisher_exact_test",
+    "deming_regression",
+    "orthogonal_regression",
     "kalman_regression_beta",
     "theil_sen",
     "median_absolute_deviation",